// src/types/appRoutes.ts
export enum AppRoute {
  Home = "/",
  SignIn = "/signin",
  SignInTeacher = "/signin/teacher",
  SignUp = "/signup",
  SignUpTeacher = "/signup/teacher",
  TeacherDashboard = "/teacher/dashboard",
  TeacherDashboardProfile = "/teacher/dashboard/profile",
  TeacherDashboardGrowthCenter = "/teacher/dashboard/growth-center",
  TeacherSchedule = "/teacher/schedule",
  TeacherMessages = "/teacher/dashboard/messages",
  TeacherStudents = "/teacher/students",
  TeacherStudio = "/teacher/studio",
  TeacherSonataSite = "/teacher/sonata-site",
  TeacherAnalytics = "/teacher/analytics",
  TeacherSettings = "/teacher/settings",
}